stm32ai-modelzoo/requirements.txt
tensorflow==2.8.3
mlflow==2.3.0
hydra-core==1.3.2
omegaconf==2.3.0
numpy==1.23.4
pandas==1.5.3
wget==3.2
munch==2.5.0
seaborn==0.12.2
scikit-learn==1.2.2
tqdm==4.65.0
skl2onnx==1.14.0
onnx==1.12.0
onnxconverter-common==1.13.0
onnxruntime==1.14.1
optuna==3.1.1
imgaug==0.4.0
matplotlib==3.5.1
opencv_python==4.6.0.66
joblib==1.2.0
librosa==0.10.0.post2
mako==1.2.4
pyserial==3.5
boto3
botocore
sagemaker==2.154.0
